ゲームの状態管理のクラスを定義、

+++ Game/Actor/Player/Player.cpp
@@ -435,7 +435,7 @@ void Player::AdjustParam() {
     if (ImGui::CollapsingHeader(groupName_.c_str())) {
         ImGui::PushID(groupName_.c_str());
 
-        ImGui::DragInt("Hp", &hp_);
+        ImGui::DragFloat("Hp", &hp_);
 
         // EditParameter
         ImGui::Separator();
